refactor: Use common AppWrapper for intents and browser targets
